Compatibility fix L5.8

+++ database/migrations/2020_12_07_112000_create_mail_templates_table.php
@@ -17,13 +17,13 @@ class CreateMailTemplatesTable extends Migration
             $table->text('text_template')->nullable();
             $table->string('code')->nullable();
             $table->string('label')->nullable();
-            
-            $table->unsignedBigInteger('type_id');
+
+            $table->unsignedInteger('type_id');
             $table->foreign('type_id')
                 ->references('id')
                 ->on('mail_template_types')
                 ->onUpdate('cascade');
-            
+
             $table->timestamps();
         });
     }
